refactor(rtx): isolate native Vulkan renderer facade
