+++ Quetoo.vs15/libs/librenderer.vcxproj
@@ -32,6 +32,7 @@
     <ClCompile Include="..\..\src\client\renderer\r_rtx_device.c" />
     <ClCompile Include="..\..\src\client\renderer\r_rtx_bridge.c" />
     <ClCompile Include="..\..\src\client\renderer\r_rtx_output.c" />
+    <ClCompile Include="..\..\src\client\renderer\r_rtx_renderer.c" />
     <ClCompile Include="..\..\src\client\renderer\r_rtx_scene.c" />
     <ClCompile Include="..\..\src\client\renderer\r_entity.c" />
     <ClCompile Include="..\..\src\client\renderer\r_image.c" />
@@ -67,6 +68,7 @@
     <ClInclude Include="..\..\src\client\renderer\r_rtx_device.h" />
     <ClInclude Include="..\..\src\client\renderer\r_rtx_bridge.h" />
     <ClInclude Include="..\..\src\client\renderer\r_rtx_output.h" />
+    <ClInclude Include="..\..\src\client\renderer\r_rtx_renderer.h" />
     <ClInclude Include="..\..\src\client\renderer\r_rtx_scene.h" />
     <ClInclude Include="..\..\src\client\renderer\r_entity.h" />
     <ClInclude Include="..\..\src\client\renderer\r_image.h" />
