+++ src/client/renderer/Makefile.am
@@ -16,6 +16,7 @@ noinst_HEADERS = \
 	r_rtx_device.h \
 	r_rtx_bridge.h \
 	r_rtx_output.h \
+	r_rtx_renderer.h \
 	r_rtx_scene.h \
 	r_entity.h \
 	r_image.h \
@@ -76,6 +77,7 @@ librenderer_la_SOURCES += \
 	r_rtx_device.c \
 	r_rtx_bridge.c \
 	r_rtx_output.c \
+	r_rtx_renderer.c \
 	r_rtx_scene.c \
 	r_rtx_probe.c
 endif
